Identifier Table design
When FuncDeclare is called, all the statements following should be
translated to assembly language for that function and stored
sequentially after the "entrance code". During this translation,
compiler has to know what each identifier means, including local
variables and value parameters. As a result, both local variables and
value parameters have to be stored in the local identifier table, which
is basically a new, empty table constructed just for this function.
After FuncDeclare returns, no other function will have access to the
local variables and value parameters anymore, so it is safe to destruct
the local identifier table before FuncDeclare returns. This applies to
VoidFuncDeclare as well.
However, the pointer to the identifier table are stored as a member
variable in GrammarDecoder. Thus it is noteworthy that in addition to
constructing and destructing the local Identifier Table, the pointer
also needs to be changed accordingly. A new pointer pointing to the
global Identifier Table has to be added and kept consistent for the
whole grammar decoding process so that global constants and variables
can be decoded and functions can be stored and called.
It is also noteworthy that no function or void function can be stored
in a local identifier table since C doesn’t allow function declaration
in another function.