Two IDT pointers design
When ConstDeclare is called, actually it doesn’t know whether the
constants belong to the local Identifier Table or the global one. Since
the static constants can only be declared at the beginning of the
source program, it is viable to set both global and local pointers to
the global Identifier Table. In this way, ConstDeclare only has to log
each constant with the local pointer. After encountering the first
function declaration, the local pointer will be changed to point to
actual local Identifier Table for a particular function or the main
function, where using local pointer to log constants in  ConstDeclare
also works.