Completely redesign member functions for Identifier Table
Now about the address part of the Identifier Table. It is viable to
just generate unique labels for global identifiers. For global
identifiers, they are stored in .data (or .bss) section. As a result,
only one instruction together with a unique label is required. In the
following grammar decoding process, an identifier can be accessed with
its label bracketed in square brackets. For local variables, both
constants and variables are stored in stack of the function.
Instructions to allocate space (for variables) and store values (for
constants) has to be generated when filling in the Identifier Table and
the local identifiers are accessed by its offset from %ebp. This makes
it necessary to differentiate between global and local Identifier Table.